Something that approximate how I imagine the demultiplexer will work.
Honestly, I'm not sure it's worth the effort. First, I'm not even sure it's
thread-safe. Each observer gets the same slice of data, who's to say one won't
mess it up for the others? Is creating a copy of each chunklet really the only
safe way to do this? If so, maybe we should limit this effort to a single client
+ the cache...